Move SurfaceSelection into its own unit as well
