+++ src/openrct2/world/map_generator/PngTerrainGenerator.cpp
@@ -18,6 +18,7 @@
 #include "../tile_element/SurfaceElement.h"
 #include "MapGen.h"
 #include "MapHelpers.h"
+#include "SurfaceSelection.h"
 
 #include <algorithm>
