+++ src/openrct2/world/map_generator/SurfaceSelection.h
@@ -1,0 +1,18 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2025 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include "../../object/ObjectTypes.h"
+
+struct MapGenSettings;
+
+ObjectEntryIndex MapGenSurfaceTextureId(MapGenSettings* settings);
+ObjectEntryIndex MapGenEdgeTextureId(MapGenSettings* settings, ObjectEntryIndex surfaceTextureId);
+ObjectEntryIndex MapGenBeachTextureId();
